requirements.txt
Flask>=2.3.0
pyaoscx>=2.0
python-dotenv>=1.0.0
requests>=2.28.0
urllib3>=1.26.0
pycentral>=0.7.0